revised scratchpaper1.py
Previous idea had too many requirements from user where an error could easily occur. This makes it easy to run from the command line and is a lot more usable for any company or ANYONE to use. This generates matrices into a scratchoffs.txt file that gets automatically replaced with each time it is ran.